refactor(core): archive old providers & controllers
