+++ src/dto/test/archive/create-section-in-test.dto.ts
@@ -1,4 +1,4 @@
-import { SectionType } from '../../config/enums/section-types.enum';
+import { SectionType } from '../../../config/enums/section-types.enum';
 
 export class CreateSectionInTestDto {
   sectionType: SectionType;
